adding the client logo with template
