+++ components/blocks/clientLogos.tsx
@@ -1,7 +1,6 @@
-import dynamic from "next/dynamic";
+import Image from "next/image";
+import { Template } from "tinacms";
 import layoutData from "../../content/global/index.json";
-
-const Image = dynamic(() => import("next/image"));
 
 const clientsData = layoutData.clients.clientsList;
 
@@ -16,11 +15,27 @@ export const ClientLogos = () => {
             alt={client.clientName + " logo"}
             height={113}
             width={200}
-            // commented out to test whether this is breaking images on the homepage see #2368
-            //sizes="20vw"
+            loading="lazy"
             className="max-w-full rounded-lg"
           />
         ))}
     </div>
   );
+};
+
+export const clientLogosBlockSchema: Template = {
+  name: "ClientLogos",
+  label: "Client Logos",
+  ui: {
+    previewSrc: "/images/thumbs/tina/client-logos.jpg",
+  },
+  // Todo: Find a way to have no fields - the one below is to satisfy compiler
+  fields: [
+    {
+      type: "string",
+      label: "Alt text",
+      name: "altText",
+      required: true,
+    },
+  ],
 };
